Factor out define.h and typedef.h
